Went back to the old way of calling SubmitInput/QueryOutput directly from the submit function. Doesn't seem to be any less performant, and we don't have to worry about multiple thread woes.
When settings change, do a full drain/terminate/create, instead of flush/re-init.
Seems to be a lot smoother, and existing output data is preserved, even when using pre-analysis/lookahead.

Improved Linux texture-based encoding.
AllocSurfaceEx actually reuses VkImages under the hood, so we use a map to cache copy command buffers.
This greatly simplifies things, and seems to be a lot less crashy.