Re-implemented and improved the ROI feature.
It wasn't working before, as AllocSurfaceEx with AMF_MEMORY_CPU_DEFAULT didn't work properly.
Changing the access mode to READ|WRITE fixed it.

We don't cache surfaces anymore as it was causing crashes.
Now, we create a ROI data buffer, which is updated only when regions change.
The buffer is then copied to each frame's ROI surface.